+++ module/ListToDo.js
@@ -1,0 +1,11 @@
+
+class ListToDo{
+    constructor(description=null, index = null,  completed = false){
+        this.description = description;
+        this.completed = completed;
+        this.index = index;
+
+    }
+}
+
+export default ListToDo
